most tests pass

+++ request.h
@@ -4,7 +4,7 @@
 #include "log.h"
 
 typedef struct Threads_stats {
-    unsigned long id; // Thread ID
+    int id; // Thread ID
     int stat_req;     // Number of static requests handled
     int dynm_req;     // Number of dynamic requests handled
     int post_req;     // Number of POST requests handled
